+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KH3PK0561G3588MF899Y0233"
+        "01KH3W7ZW24720B399NSAWSA3C"
     )
     r"""The unique identifier of the data source"""