+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KH3PK0591EF7D9X4CRGTQETK"
+        "01KH3W7ZW5Y1Q9PJQPRDXG73RM"
     )
     r"""The unique identifier of the data source"""