+++ packages/orq-rc/RELEASES.md
@@ -5459,3 +5459,13 @@ Based on:
 - [python v4.3.0-rc.46] packages/orq-rc
 ### Releases
 - [PyPI v4.3.0-rc.46] https://pypi.org/project/orq-ai-sdk/4.3.0-rc.46 - packages/orq-rc
+
+## 2026-02-10 13:38:20
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.709.3 (2.814.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.3.0-rc.47] packages/orq-rc
+### Releases
+- [PyPI v4.3.0-rc.47] https://pypi.org/project/orq-ai-sdk/4.3.0-rc.47 - packages/orq-rc